+++ app/src/main/java/org/synapseworks/pageharbor/ui/home/HomeScreen.kt
@@ -102,7 +102,7 @@ fun HomeScreen(
                 ) {
                     Text(
                         modifier = Modifier.semantics { heading() },
-                        text = stringResource(R.string.app_name),
+                        text = stringResource(R.string.app_name_short),
                         style = MaterialTheme.typography.titleLarge,
                         color = MaterialTheme.colorScheme.onBackground,
                         textAlign = TextAlign.Center,